+++ src/issuingTransaction/issuingTransaction.php
@@ -63,12 +63,12 @@ class IssuingTransaction extends Resource
     Receive an enumerator of IssuingTransaction objects previously created in the Stark Infra API
 
     ## Parameters (optional):
-        - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
-        - externalIds [list of strings, default []]: external IDs. ex: ["5656565656565656", "4545454545454545"]
+        - tags [array of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+        - externalIds [array of strings, default []]: external IDs. ex: ["5656565656565656", "4545454545454545"]
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
         - status [string, default null]: filter for status of retrieved objects. ex: "approved", "canceled", "denied", "confirmed" or "voided"
-        - ids [list of strings, default [], default null]: purchase IDs
+        - ids [array of strings, default [], default null]: purchase IDs
         - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
@@ -90,12 +90,12 @@ class IssuingTransaction extends Resource
 
     ## Parameters (optional):
         - cursor [string, default null]: cursor returned on the previous page function call
-        - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
-        - externalIds [list of strings, default []]: external IDs. ex: ["5656565656565656", "4545454545454545"]
+        - tags [array of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+        - externalIds [array of strings, default []]: external IDs. ex: ["5656565656565656", "4545454545454545"]
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
         - status [string, default null]: filter for status of retrieved objects. ex: "approved", "canceled", "denied", "confirmed" or "voided"
-        - ids [list of strings, default [], default null]: purchase IDs
+        - ids [array of strings, default [], default null]: purchase IDs
         - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call