+++ src/pixChargeback/log.php
@@ -21,7 +21,7 @@ class Log extends Resource
         - id [string]: unique id returned when the log is created. ex: "5656565656565656"
         - created [string]: creation datetime for the log. 
         - type [string]: type of the PixChargeback event which triggered the log creation. 
-        - errors [list of strings]: list of errors linked to this PixChargeback event
+        - errors [array of strings]: list of errors linked to this PixChargeback event
         - chargeback [PixChargeback]: PixChargeback entity to which the log refers to.
     */
     function __construct(array $params)
@@ -61,12 +61,12 @@ class Log extends Resource
     Receive an enumerator of PixChargeback\Log objects previously created in the Stark Infra API
     
     ## Parameters (optional):
-        - ids [list of strings, default null]: Log ids to filter PixChargeback Logs. ex: ["5656565656565656"]
+        - ids [array of strings, default null]: Log ids to filter PixChargeback Logs. ex: ["5656565656565656"]
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
-        - chargebackIds [list of strings, default null]: list of PixChargeback IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [array of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
+        - chargebackIds [array of strings, default null]: list of PixChargeback IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call
     
     ## Return:
@@ -88,12 +88,12 @@ class Log extends Resource
     
     ## Parameters (optional):
         - cursor [string, default null]: cursor returned on the previous page function call
-        - ids [list of strings, default null]: Log ids to filter PixChargeback Logs. ex: ["5656565656565656"]
+        - ids [array of strings, default null]: Log ids to filter PixChargeback Logs. ex: ["5656565656565656"]
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
-        - chargebackIds [list of strings, default null]: list of PixChargeback IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [array of strings, default null]: filter retrieved objects by types. ex: "created", "failed", "delivering", "delivered", "closed", "canceled".
+        - chargebackIds [array of strings, default null]: list of PixChargeback IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call
     
     ## Return: