+++ src/pixKey/log.php
@@ -22,7 +22,7 @@ class Log extends Resource
         - id [string]: unique id returned when the log is created. ex: "5656565656565656"
         - created [Date, Datetime or string]: creation datetime for the log. ex: "2020-03-10 10:30:00.000"
         - type [string]: type of the PixKey event which triggered the log creation. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
-        - errors [list of strings]: list of errors linked to this PixKey event
+        - errors [array of strings]: list of errors linked to this PixKey event
         - key [PixKey]: PixKey entity to which the log refers to.
     */    
     function __construct(array $params)
@@ -62,12 +62,12 @@ class Log extends Resource
     Receive an enumerator of PixKey\Log objects previously created in the Stark Infra API
     
     ## Parameters (optional):
-        - ids [list of strings, default null]: Log ids to filter PixKey Logs. ex: ["5656565656565656"]
+        - ids [array of strings, default null]: Log ids to filter PixKey Logs. ex: ["5656565656565656"]
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "created","registered","updated","failed","canceling" and "canceled".
-        - keyIds [list of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [array of strings, default null]: filter retrieved objects by types. ex: "created","registered","updated","failed","canceling" and "canceled".
+        - keyIds [array of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call
     
     ## Return:
@@ -88,12 +88,12 @@ class Log extends Resource
     
     ## Parameters (optional):
         - cursor [string, default null]: cursor returned on the previous page function call
-        - ids [list of strings, default null]: Log ids to filter PixKey Logs. ex: ["5656565656565656"]
+        - ids [array of strings, default null]: Log ids to filter PixKey Logs. ex: ["5656565656565656"]
         - limit [integer, default 100]: maximum number of objects to be retrieved. Max = 100. ex: 35
         - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
         - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
-        - types [list of strings, default null]: filter retrieved objects by types. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
-        - keyIds [list of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - types [array of strings, default null]: filter retrieved objects by types. ex: "created", "registered", "updated", "failed", "canceling" and "canceled".
+        - keyIds [array of strings, default null]: list of PixKey IDs to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was set before function call
     
     ## Return: